access log filters: various enhancements (#2588)
Implement LE comparison filter.
Runtime filter additions (configurable default, divisor, and independent
randomness).
Risk Level: Low
Testing: Unit
Docs Changes: envoyproxy/data-plane-api#490
API Changes: envoyproxy/data-plane-api#490
Release Notes: Added
Fixes #2570

+++ include/envoy/runtime/runtime.h
@@ -92,11 +92,11 @@ public:
    * @param random_value supplies the stable random value to use for determining whether the feature
    *        is enabled.
    * @param control max number of buckets for sampling. Sampled value will be in a range of
-   * [0, num_buckets).
+   *        [0, num_buckets).
    * @return true if the feature is enabled.
    */
   virtual bool featureEnabled(const std::string& key, uint64_t default_value, uint64_t random_value,
-                              uint16_t num_buckets) const PURE;
+                              uint64_t num_buckets) const PURE;
 
   /**
    * Fetch raw runtime data based on key.
